% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_dataInput_ui}
\alias{mod_dataInput_ui}
\alias{mod_dataInput_server}
\title{Module to select dataset for visualization}
\usage{
mod_dataInput_ui(id)

mod_dataInput_server(input, output, session, parentSession)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Gives functioning to upload, download dataset for visualization. User can also choose existing datasets
}
\keyword{internal}
