% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_missing_data.R
\name{mod_missing_data_ui}
\alias{mod_missing_data_ui}
\alias{mod_missing_data_server}
\title{mod_missing_data_ui and mod_missing_data_server}
\usage{
mod_missing_data_ui(id)

mod_missing_data_server(input, output, session, dataset_missing)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
