% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxonomic.R
\name{mod_taxonomic_ui}
\alias{mod_taxonomic_ui}
\alias{mod_taxonomic_server}
\title{Module to visualize taxonomic data}
\usage{
mod_taxonomic_ui(id)

mod_taxonomic_server(input, output, session, data_taxo)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Contain plots to visulaize taxonomic related data fields.
}
\keyword{internal}
