% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_temporal.R
\name{mod_temporal_ui}
\alias{mod_temporal_ui}
\alias{mod_temporal_server}
\title{Module to visualize temporal data}
\usage{
mod_temporal_ui(id)

mod_temporal_server(input, output, session, data_temporal)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
This module is for visualization of time related data.
}
\keyword{internal}
