% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataSummary.R
\name{mod_dataSummary_ui}
\alias{mod_dataSummary_ui}
\alias{mod_dataSummary_server}
\title{mod_dataSummary_ui and mod_dataSummary_server}
\usage{
mod_dataSummary_ui(id)

mod_dataSummary_server(input, output, session, dataset)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
