% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spatial.R
\name{mod_spatial_ui}
\alias{mod_spatial_ui}
\alias{mod_spatial_server}
\title{mod_spatial_ui and mod_spatial_server}
\usage{
mod_spatial_ui(id)

mod_spatial_server(input, output, session, data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
