% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxonomic.R
\name{mod_taxonomic_ui}
\alias{mod_taxonomic_ui}
\alias{mod_taxonomic_server}
\title{mod_taxonomic_ui and mod_taxonomic_server}
\usage{
mod_taxonomic_ui(id)

mod_taxonomic_server(input, output, session, dataTaxo)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
