% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_temporal.R
\name{mod_temporal_ui}
\alias{mod_temporal_ui}
\alias{mod_temporal_server}
\title{mod_temporal_ui and mod_temporal_server}
\usage{
mod_temporal_ui(id)

mod_temporal_server(input, output, session, dataTemporal)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
